Extract sshTunnel - kic tunnel
