+++ pkg/minikube/tunnel/kic/tunnel.go
@@ -1,8 +1,6 @@
 package kic
 
 import (
-	"fmt"
-	"os/exec"
 	"time"
 
 	v1 "k8s.io/api/core/v1"
@@ -15,12 +13,6 @@ type Tunnel struct {
 	sshPort string
 	sshKey  string
 	v1Core  typed_core.CoreV1Interface
-}
-
-type sshTunnel struct {
-	name string
-	kill bool
-	cmd  *exec.Cmd
 }
 
 // NewTunnel ...
@@ -43,7 +35,7 @@ func (t *Tunnel) Start() error {
 		}
 
 		for _, v := range sshTunnels {
-			v.kill = true
+			v.sigkill = true
 		}
 
 		for _, s := range services.Items {
@@ -51,17 +43,17 @@ func (t *Tunnel) Start() error {
 				sshTunnel, ok := sshTunnels[s.Name]
 
 				if ok {
-					sshTunnel.kill = false
+					sshTunnel.sigkill = false
 					continue
 				}
 
-				newSSHTunnel := t.createSSHTunnel(s.Name, s.Spec.ClusterIP, s.Spec.Ports)
+				newSSHTunnel := createSSHTunnel(t, s.Name, s.Spec.ClusterIP, s.Spec.Ports)
 				sshTunnels[newSSHTunnel.name] = newSSHTunnel
 			}
 		}
 
 		for _, v := range sshTunnels {
-			if v.kill {
+			if v.sigkill {
 				v.stop()
 				delete(sshTunnels, v.name)
 			}
@@ -69,61 +61,5 @@ func (t *Tunnel) Start() error {
 
 		// which time to use?
 		time.Sleep(1 * time.Second)
-	}
-}
-
-func (t *Tunnel) createSSHTunnel(name, clusterIP string, ports []v1.ServicePort) *sshTunnel {
-	// extract sshArgs
-	sshArgs := []string{
-		"-N",
-		"docker@127.0.0.1",
-		"-p", t.sshPort,
-		"-i", t.sshKey,
-	}
-
-	for _, port := range ports {
-		arg := fmt.Sprintf(
-			"-L %d:%s:%d",
-			port.Port,
-			clusterIP,
-			port.Port,
-		)
-
-		sshArgs = append(sshArgs, arg)
-	}
-
-	cmd := exec.Command("ssh", sshArgs...)
-
-	// TODO: name must be different, because if a service was changed,
-	// we must remove the old process and create the new one
-	s := &sshTunnel{
-		name: fmt.Sprintf("%s", name),
-		kill: false,
-		cmd:  cmd,
-	}
-
-	go s.run()
-
-	return s
-}
-
-func (s *sshTunnel) run() {
-	fmt.Println("running", s.name)
-	err := s.cmd.Start()
-	if err != nil {
-		// TODO: improve logging
-		fmt.Println(err)
-	}
-
-	// we are ignoring wait return, because the process will be killed, once the tunnel is not needed.
-	s.cmd.Wait()
-}
-
-func (s *sshTunnel) stop() {
-	fmt.Println("stopping", s.name)
-	err := s.cmd.Process.Kill()
-	if err != nil {
-		// TODO: improve logging
-		fmt.Println(err)
 	}
 }
